Moved SevenSegmentDisplay to hat.util.ui (so we can use it in other examples)
